% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{My hello world function}
\usage{
hello(n = 5)
}
\arguments{
\item{n}{Number of random numbers}
}
\value{
A vector of length n
}
\description{
My hello world function
}
\examples{
res <- hello(5)
\dontrun{
hello(10)
}
}
